\name{fitsoilwater_App}
\alias{fitsoilwater_App}

\title{Fitting soil water retention curves}

\description{A shiny for fitting soil water retention curves}

\usage{
  fitsoilwater_App()
}

\arguments{
  \item{}{Available on the interactive web page}
}

\details{Run online from: https://appsoilphysics.shinyapps.io/fitsoilwaterAPP/}

\value{shiny app}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }
