\name{LLWR_App}
\alias{LLWR_App}

\title{Least limiting water range}

\description{A shiny for calculation of the usual least limiting water range}

\usage{
  LLWR_App()
}

\arguments{
  \item{}{Available on the interactive web page}
}

\details{Run online from: https://soilphysics.shinyapps.io/LLWRAPP/}

\value{shiny app}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }
