\name{r_App}
\alias{r_App}

\title{Capillarity equation}

\description{A shiny for equation of capillarity}

\usage{
  r_App()
}

\arguments{
  \item{}{Available on the interactive web page}
}


\details{Run online from: https://soilphysics.shinyapps.io/capilarAPP/}

\value{shiny app}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }

