\name{IndSoilWater_App}
\alias{IndSoilWater_App}

\title{Exploring the water retention curve}

\description{A shiny for exploration of the water retention curve}

\usage{
  IndSoilWater_App()
}

\arguments{
  \item{}{Available on the interactive web page}
}

\details{Run online from: https://soilphysics.shinyapps.io/wrcAPP/}

\value{shiny app}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }

