\name{LLWR_LLMPR_App}
\alias{LLWR_LLMPR_App}

\title{Least limiting water and matric potential ranges of agricultural soils}

\description{A shiny for calculation of least limiting water and matric potential ranges of agricultural soils with calculated physical restriction thresholds}

\usage{
  LLWR_LLMPR_App()
}

\arguments{
  \item{}{Available on the interactive web page}
}

\details{Run online from: https://soilphysics.shinyapps.io/LLWR_LLMPR/}

\value{shiny app}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }
